Refactor to RawBatch and CleanBatch wrapper types
